requirements.txt
Django==4.2.7
pandas
pypdf
python-docx
python-pptx
python-decouple==3.8
Pillow>=10.0.0

mammoth
openpyxl
docxcompose
